MemoDataSourceを一つのファイルでまとめる
