--- app/src/main/java/jp/co/cyberagent/dojo2020/data/MemoDataSource.kt
@@ -1,14 +1,0 @@
-package jp.co.cyberagent.dojo2020.data
-
-import jp.co.cyberagent.dojo2020.data.model.Memo
-import kotlinx.coroutines.flow.Flow
-
-interface MemoDataSource {
-    suspend fun saveMemo(memo: Memo)
-
-    suspend fun fetchAllMemo(): Flow<List<Memo>>
-
-    suspend fun fetchMemoById(id: String): Flow<Memo?>
-
-    suspend fun deleteMemoById(id: String)
-}
