+++ app/src/main/java/jp/co/cyberagent/dojo2020/data/MemoRepository.kt
@@ -1,5 +1,6 @@
 package jp.co.cyberagent.dojo2020.data
 
+import jp.co.cyberagent.dojo2020.data.local.MemoDataSource
 import jp.co.cyberagent.dojo2020.data.model.Memo
 import jp.co.cyberagent.dojo2020.data.remote.firestore.FireStoreDataSource
 import kotlinx.coroutines.FlowPreview
