+++ app/src/main/java/jp/co/cyberagent/dojo2020/data/local/DefaultMemoDataSource.kt
@@ -1,13 +1,22 @@
 package jp.co.cyberagent.dojo2020.data.local
 
-import jp.co.cyberagent.dojo2020.data.MemoDataSource
 import jp.co.cyberagent.dojo2020.data.local.db.ApplicationDataBase
 import jp.co.cyberagent.dojo2020.data.local.db.memo.MemoEntity
 import jp.co.cyberagent.dojo2020.data.model.Memo
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.map
 
-class LocalMemoDataSource(private val dataBase: ApplicationDataBase) : MemoDataSource {
+interface MemoDataSource {
+    suspend fun saveMemo(memo: Memo)
+
+    suspend fun fetchAllMemo(): Flow<List<Memo>>
+
+    suspend fun fetchMemoById(id: String): Flow<Memo?>
+
+    suspend fun deleteMemoById(id: String)
+}
+
+class DefaultMemoDataSource(private val dataBase: ApplicationDataBase) : MemoDataSource {
 
     override suspend fun saveMemo(memo: Memo) {
         val forInsertMemo = MemoEntity(memo.id, memo.title, memo.contents, memo.time)
